Unify nunit-console and nunitlite.runner option processing
Unify the command line/option processing between nunit-console and
nunit.runner.
Also, add the --testlist option to nunit.runner (and eventually,
nunit-console) to allow a file containing a list of tests to be run to
be specified.

Unify the tests for both as well, and add tests for the --testfile
option.