Separate `RangeAttribute` from 'ValuesAttribute`
Previously, `RangeAttribute` derived from `ValuesAttribute`. However,
there are certain use cases which apply to the `RangeAttribute`, but not
to the `ValuesAttribute`. For example, we may want specify multiple,
distinct ranges for a parameter by adding the `RangeAttribute` multiple
times, but there isn't really a corresponding use case for the values
attribute. We could just specify all the desired values in on attribute.

This change updates the `RangeAttribute` to no longer derive from the
`ValuesAttribute`. Instead, it now derives from `DataAttribute` and
implements `IParameterDataSource`. The `IParameterDataSource.GetData`
method has been implemented in `RangeAttribute` to preserve the behavior
of that method in the `ValuesAttribute`. However, certain cases that
could not be hit in the `RangeAttribute` (e.g. when the internal data
was non-numeric) were removed.

Fixes #2469