Allow null values in CollectionOrderedConstraint
Allows null values for simple ordering in CollectionOrderedConstraint.
Null sorts less than everything else. An ArgumentNullException will
still be thrown if ordering by a property.

This is a potentially breaking change. If the previous behavior of
throwing an exception if a collection contains null items is desired,
it can be achieved using `Has.None.Null.And.Ordered`.

Fixes #1473