ConstraintExpression.AnyOf returns AnyOfConstraint
Other methods of ConstraintExpression (e.g. SubPathOf, SamePath, Contains)
have a return signature of their specific constraint (SubPathConstraint,
SamePathConstraint, ContainsConstraint or SomeItemsConstraint). For AnyOf
the signature was just Constraint. Changed the signature to returning
AnyOfConstraint and in the return casted the response from Append like
it was done in other constraint.
    
Addresses Issue #3024