Improve performance of discovery/execution #2488
Changes the regex that stripped out invalid XML characters to an equivalent method that will not allocate more strings unless there is an invalid character in the string. This was a performance issue for me when attempting to run a single test in a project with 20,000 tests with the VS test runner as this method took nearly 8s before (now down to < 1s with my changes).