Compare element within ValueTuples with NUnitEqualityComparer
ValueTuples implement IEquatable and we tried to compare using
EquatablesComparer before ValueTupleComparer, so ValueTuples
was compared using the standard semantics and not using
NUnitEqualityComparer, so e.g. Dictionaries was compared using
reference equality and not using DictionariesComparer.

Note that the test for Tuples passed before this change
as Tuples does not implement IEquatable.

Fixes #3073