Enumerate enumerables to determine emptiness
The usage of the `Count` property can lead to unexpected results
if the `Count` is not in sync with the enumeration.

It makes sense to use the `Count` on a `ICollection` since the interface
clearly connects the `Count` with the enumeration. On any `IEnumerable`,
this assumption is not always correct though.