Enable more flexible tolerance use.
Tolerance has been silently ignored in cases where certain
types of equality were requested.  It was also impossible
for someone to implement an External Comparer that considered
tolerance.  This enables custom EqualityAdapters to process
tolerances.

Tolerance gained the concept of HasVariance to better determine
if strict equals could be applied or not.  Within() calls that
were previously ignored can now determine if it's possible to
apply the tolerance, and if not an error is raised.