Add support for NET8.0 as build target
Main item it to drop SerializationConstructor from Exception classes

From the recommended actions in https://github.com/dotnet/docs/issues/34893:

Consider whether you truly need your custom exception type to be serializable. Chances are you do not need it to be serializable, as exception serialization is primarily intended to support remoting, and support for remoting was dropped in .NET Core 1.0.